FEAT - Implement the buddies section
In this section, each user can create a Code Buddy request to work
together in a project on github.
An user can scroll the buddies request list and find an interesting
project to contribute and then, click in the reporsitory link directly
to the buddy respository and start contributing.
The intention is to promote user to contribute with other peoples
projects.